+++ src/main/java/shopping/auth/util/BearerExtractor.java
@@ -4,7 +4,7 @@ import javax.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
-import shopping.auth.exception.ShoppingAuthenticationException;
+import shopping.global.exception.ShoppingAuthenticationException;
 
 @Component
 public class BearerExtractor {
