+++ src/main/java/shopping/global/exception/GlobalExceptionHandler.java
@@ -6,7 +6,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import shopping.auth.exception.ShoppingAuthenticationException;
 import shopping.global.exception.dto.ExceptionResponse;
 
 @ControllerAdvice
