+++ src/main/java/shopping/global/exception/ShoppingAuthenticationException.java
@@ -1,6 +1,4 @@
-package shopping.auth.exception;
-
-import shopping.global.exception.ShoppingException;
+package shopping.global.exception;
 
 public class ShoppingAuthenticationException extends ShoppingException {
