+++ src/test/java/shopping/auth/BearerExtractorTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.mock.web.MockHttpServletRequest;
-import shopping.auth.exception.ShoppingAuthenticationException;
+import shopping.global.exception.ShoppingAuthenticationException;
 import shopping.auth.util.BearerExtractor;
 
 @DisplayName("Bearer 토큰 관련 기능 테스트")
