+++ src/test/java/shopping/domain/order/OrderTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import shopping.exchange.MockExchangeRateApi;
 import shopping.global.vo.Name;
-import shopping.cart.domain.vo.Price;
+import shopping.global.vo.Price;
 import shopping.global.vo.Quantity;
 import shopping.order.domain.Order;
 import shopping.order.domain.OrderProduct;
